Fix the isLocked precompile so it can work
