+++ crosschain/src/main/java/org/hyperledger/besu/crosschain/ethereum/privatenet/precompiles/CrosschainPrecompiledContractRegistries.java
@@ -41,7 +41,7 @@ public class CrosschainPrecompiledContractRegistries {
     registry.put(
         Address.CROSSCHAIN_ISLOCKED,
         Account.DEFAULT_VERSION,
-        new CrosschainGetInfoPrecompiledContract(
+        new CrosschainIsLockedPrecompiledContract(
             precompiledContractConfiguration.getGasCalculator()));
     return registry;
   }
